Add caching of next item in generic heap
From Juri's email:
"I'm using the [binomial heap] code to create a simulator of the Linux
scheduler behavior in userspace. [...] Since I want to mimic the Linux
push/pull mechanism I need a way to cache min and next active tasks on
each runqueue. I decided to modify your code to insert a next cache
near the "min" one. [...] I attached patch that introduces the "next"
cache and a different test for it."

[ Committer's note: slightly edited for naming, to be -Werror
  compliant, and to remove Linux-specific headers. ]